Add Dsl setting and split out oauth
